+++ tests/gui/actions/ActionTests.cpp
@@ -39,6 +39,7 @@ void ActionTests::EqualsOperator_Should_ReturnTrue_When_AllValuesMatch()
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -53,6 +54,7 @@ void ActionTests::EqualsOperator_Should_ReturnTrue_When_AllValuesMatch()
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -77,6 +79,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_NameIsDifferent()
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName("Other");
@@ -91,6 +94,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_NameIsDifferent()
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -115,6 +119,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_ShortcutDifferent()
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -129,6 +134,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_ShortcutDifferent()
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -153,6 +159,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsCaptureEnabledDiffere
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -167,6 +174,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsCaptureEnabledDiffere
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -191,6 +199,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IncludeCursorDifferent(
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -205,6 +214,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IncludeCursorDifferent(
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -229,6 +239,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_CaptureDelayDifferent()
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -243,6 +254,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_CaptureDelayDifferent()
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -267,6 +279,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_CaptureModeDifferent()
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -281,6 +294,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_CaptureModeDifferent()
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -305,6 +319,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsSaveEnabledDifferent(
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -319,6 +334,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsSaveEnabledDifferent(
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -343,6 +359,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsCopyToClipboardEnable
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -357,6 +374,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsCopyToClipboardEnable
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -381,6 +399,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsUploadEnabledDifferen
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -395,6 +414,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsUploadEnabledDifferen
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -419,6 +439,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsOpenDirectoryEnabledD
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -433,6 +454,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsOpenDirectoryEnabledD
 	action2.setIsOpenDirectoryEnabled(false);
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -457,6 +479,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsPinScreenshotEnabledD
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -471,6 +494,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsPinScreenshotEnabledD
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(false);
 	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
@@ -495,6 +519,7 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsHideMainWindowEnabled
 	action1.setIsOpenDirectoryEnabled(true);
 	action1.setIsPinImageEnabled(true);
 	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
 
 	Action action2;
 	action2.setName(action1.name());
@@ -509,12 +534,53 @@ void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsHideMainWindowEnabled
 	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
 	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
 	action2.setIsHideMainWindowEnabled(false);
+	action2.setIsGlobalShortcut(action1.isGlobalShortcut());
 
 	// act
 	auto result = action1 == action2;
 
 	// assert
     QCOMPARE(result, false);
+}
+
+void ActionTests::EqualsOperator_Should_ReturnFalse_When_IsGlobalShortcutDifferent()
+{
+	// arrange
+	Action action1;
+	action1.setName("Name");
+	action1.setShortcut(Qt::CTRL + Qt::Key_A);
+	action1.setIsCaptureEnabled(true);
+	action1.setIncludeCursor(true);
+	action1.setCaptureDelay(2000);
+	action1.setCaptureMode(CaptureModes::FullScreen);
+	action1.setIsSaveEnabled(true);
+	action1.setIsCopyToClipboardEnabled(true);
+	action1.setIsUploadEnabled(true);
+	action1.setIsOpenDirectoryEnabled(true);
+	action1.setIsPinImageEnabled(true);
+	action1.setIsHideMainWindowEnabled(true);
+	action1.setIsGlobalShortcut(true);
+
+	Action action2;
+	action2.setName(action1.name());
+	action2.setShortcut(action1.shortcut());
+	action2.setIsCaptureEnabled(action1.isCaptureEnabled());
+	action2.setIncludeCursor(action1.includeCursor());
+	action2.setCaptureDelay(action1.captureDelay());
+	action2.setCaptureMode(action1.captureMode());
+	action2.setIsSaveEnabled(action1.isSaveEnabled());
+	action2.setIsCopyToClipboardEnabled(action1.isCopyToClipboardEnabled());
+	action2.setIsUploadEnabled(action1.isUploadEnabled());
+	action2.setIsOpenDirectoryEnabled(action1.isOpenDirectoryEnabled());
+	action2.setIsPinImageEnabled(action1.isPinImageEnabled());
+	action2.setIsHideMainWindowEnabled(action1.isHideMainWindowEnabled());
+	action2.setIsGlobalShortcut(false);
+
+	// act
+	auto result = action1 == action2;
+
+	// assert
+	QCOMPARE(result, false);
 }
 
 TEST_MAIN(ActionTests)